# make global search case-insensitive, and allow the search to run on all locales