--- a/packages/admin/src/Resources/Resource.php
+++ b/packages/admin/src/Resources/Resource.php
@@ -445,15 +445,14 @@ class Resource
             $query->when(
                 method_exists($model, 'isTranslatableAttribute') && $model->isTranslatableAttribute($searchAttribute),
                 function (Builder $query) use ($databaseConnection, $searchAttribute, $searchOperator, $searchQuery, $whereClause): Builder {
-                    $activeLocale = app()->getLocale();
 
                     $searchColumn = match ($databaseConnection->getDriverName()) {
-                        'pgsql' => "{$searchAttribute}->>'{$activeLocale}'",
-                        default => "json_extract({$searchAttribute}, \"$.{$activeLocale}\")",
+                        'pgsql' => "{$searchAttribute}::text",
+                        default => "json_extract({$searchAttribute}, '$')",
                     };
 
                     return $query->{"{$whereClause}Raw"}(
-                        "lower({$searchColumn}) {$searchOperator} ?",
+                        "lower({$searchColumn}) {$searchOperator} lower(?)",
                         "%{$searchQuery}%",
                     );
                 },