--- a/packages/tables/src/Columns/Concerns/InteractsWithTableQuery.php
+++ b/packages/tables/src/Columns/Concerns/InteractsWithTableQuery.php
@@ -87,15 +87,14 @@ trait InteractsWithTableQuery
             $query->when(
                 method_exists($model, 'isTranslatableAttribute') && $model->isTranslatableAttribute($searchColumnName),
                 function (Builder $query) use ($searchColumnName, $searchOperator, $search, $whereClause, $databaseConnection): Builder {
-                    $activeLocale = $this->getLivewire()->getActiveTableLocale() ?: app()->getLocale();
 
                     $searchColumn = match ($databaseConnection->getDriverName()) {
-                        'pgsql' => "{$searchColumnName}->>'{$activeLocale}'",
-                        default => "json_extract({$searchColumnName}, \"$.{$activeLocale}\")",
+                        'pgsql' => "{$searchColumnName}::text",
+                        default => "json_extract({$searchColumnName}, '$')",
                     };
 
                     return $query->{"{$whereClause}Raw"}(
-                        "lower({$searchColumn}) {$searchOperator} ?",
+                        "lower({$searchColumn}) {$searchOperator} lower(?)",
                         "%{$search}%",
                     );
                 },