--- a/packages/tables/src/Columns/Concerns/InteractsWithTableQuery.php
+++ b/packages/tables/src/Columns/Concerns/InteractsWithTableQuery.php
@@ -87,14 +87,15 @@ trait InteractsWithTableQuery
             $query->when(
                 method_exists($model, 'isTranslatableAttribute') && $model->isTranslatableAttribute($searchColumnName),
                 function (Builder $query) use ($searchColumnName, $searchOperator, $search, $whereClause, $databaseConnection): Builder {
-
+                    $activeLocale = $this->getLivewire()->getActiveTableLocale() ?: app()->getLocale();
+                    
                     $searchColumn = match ($databaseConnection->getDriverName()) {
-                        'pgsql' => "{$searchColumnName}::text",
-                        default => "json_extract({$searchColumnName}, '$')",
+                        'pgsql' => "{$searchColumnName}->>'{$activeLocale}'",
+                         default => "json_extract({$searchColumnName}, \"$.{$activeLocale}\")",
                     };
 
                     return $query->{"{$whereClause}Raw"}(
-                        "lower({$searchColumn}) {$searchOperator} lower(?)",
+                        "lower({$searchColumn}) {$searchOperator} ?",
                         "%{$search}%",
                     );
                 },