fixes https://github.com/filamentphp/filament/discussions/4828