@AbanoubNassem 2022-11-04
just noticed , it broke the local `Resource` search , working on fixing that