@AbanoubNassem 2022-11-04
just noticed that the local `Resource` search , has same issue, working on fixing that as well.