@danharrin 2022-11-05
I've reverted the fix for tables because they should have a LocaleSwitcher anyway