@AbanoubNassem 2022-11-05
> I've reverted the fix for tables because they should have a LocaleSwitcher anyway

we still have the same issues as there : https://github.com/filamentphp/filament/discussions/4828 , it's case sensitive as well. 

I believe, if we have a website with like 6 different locales , to keep switching the table locale to search, is not user friendly. Imagine a team of a company working together.

for individual/single owner websites, I see that they can switch the locale easily, because they know what they are doing , but in the scale of a company , this will not be the required behaviour.

I know this has its impact on performance , therefore if you are interested , I can work on implementing `scout` search capabilities, so we get the benefits of the indexed search.