@danharrin 2022-11-05
Scout can be integrated like this - https://filamentphp.com/docs/2.x/tables/getting-started#searching-records-with-laravel-scout