@danharrin 2022-11-05
I kinda agree now with the multi-locale search... what is the performance impact for searching 100,000 records?