@AbanoubNassem 2022-11-05
> 

@danharrin 

with 110600 records,  app in debug mode with `debugbar` attached , not optimised app, on Macbook Pro 2016:

```
//current implementation 

1 character (a) : 709ms
2 character (se) : 822ms
5 character (volup) : 810ms
2 words (Amsterdam Padova) : 905ms

-----------------------------------
// with the one I provided

1 character (a) : 660μs
2 character (se) : 1.34s
5 character (volup) : 1.37s
2 words (Amsterdam Padova) : 1.61s
```

so form what I see , it's not a huge impact , it works fairly well actually.