@danharrin 2022-11-05
@zepfietje wdyt about these performance stats