@zepfietje 2022-11-05
Hmm, 180% duration is not something we should ignore. What's causing the slowdown specifically? The multi-language search?