@zepfietje 2022-11-05
Hard one. How would you toggle between locales for the global search if it didn't search for all locales by default?