@AbanoubNassem 2022-11-05
> Hard one. How would you toggle between locales for the global search if it didn't search for all locales by default?

@danharrin  @zepfietje 

I noticed something weird tho, globale search , has no impact , performance is

```
1 character (a) : 294ms
2 character (se) : 369ms
5 character (volup) : 340ms
```

while I'm searching multiple tables , not one table like with resource-table search .

not really sure , why is that!?