@AbanoubNassem 2022-11-05
> Thanks for clarifying @AbanoubNassem!
> 
> Maybe @danharrin could explain, since I haven't looked too deep into this part.
> 
> As for my question, there's a toggle above the table then, right?

@zepfietje 

no there is not , it searches on the `$activeLocale = app()->getLocale();` , so it forces the app locale , no way to toggle/change it.