@danharrin 2022-11-05
Thats only for global search as far as I remember? Tables search respects the LocaleSwitcher