@AbanoubNassem 2022-11-05
> Thats only for global search as far as I remember? Tables search respects the LocaleSwitcher

it does respect the `LocaleSwitcher` but currently it lowers the column , not the input.

and on refresh I loses the LocaleSwitcher state, moreover I have 4 locales , to keep switching the locale to find the right locale which has the input is a bit cumbersome.