@danharrin 2022-11-06
We could make a function to store the locale in the session?