@AbanoubNassem 2022-11-06
> We could make a function to store the locale in the session?

hummm how about , if we make searching all locales optional? 
so by default it's false , but if you need it you can toggle it.

` public function applySearchConstraint(Builder $query, string $search, bool &$isFirst, bool $isIndividual = false, bool $searchAllLocales = false){}`

I think this would be a better approach, what do you think @danharrin  @zepfietje ?