@danharrin 2022-11-06
You can do it already by just overriding that method on the Livewire class. Extract it into a trait in your project if you want?