@AbanoubNassem 2022-11-06
> You can do it already by just overriding that method on the Livewire class. Extract it into a trait in your project if you want?

I can't , this is done on the `Column` -> `InteractsWithTableQuery`, I think you still confusing the `Column` search with `Global Search` , global search is working perfectly fine , it's the `Column` search now which has the issue.