@danharrin 2022-11-06
Oh right, you can override it using `->searchable(query: fn (Builder $query) => ...)` then